Code/requirements.txt
ecdsa-0.19.1 
six-1.17.0 
numpy-1.23.5
scipy-1.10.1
scikit-learn-1.2.2